fixed modal for product submit, now beer
